requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.0.354
langchain-community==0.0.7
langchain-google-genai==0.0.2
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1